+++ CHANGELOG.md
@@ -9,8 +9,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Here we write upgrading notes for brands. It's a team effort to make them as straightforward as possible.
 
-### Deprecated 
-- IErrorMessage interface will be removed
+## [0.1.6-beta] - 2022-06-20
+### Added
+- [7828a80](https://github.com/KarthikUdyawar/url-shortener/commit/7828a80) - Create IResponse.ts
+- [876c368](https://github.com/KarthikUdyawar/url-shortener/commit/876c368) - Create IMessage.ts
+### Changed
+- [1e3a219](https://github.com/KarthikUdyawar/url-shortener/commit/1e3a219) - Change test.controller -> Test.controller
+### Deprecated
+- [c236c2b](https://github.com/KarthikUdyawar/url-shortener/commit/c236c2b) - IErrorMessage interface will be removed
+### Removed
+- [7c1fbba](https://github.com/KarthikUdyawar/url-shortener/commit/7c1fbba) - Delete ICreateReqBody interface
+### Fixed
+- [836904e](https://github.com/KarthikUdyawar/url-shortener/commit/836904e) - Using IResponse & IMessage interfaces in response
 
 ## [0.1.5-beta] - 2022-06-20
 ### Removed